Everything but the actual data access
- Created an interface for a session store, specifying the actions that
are required
- Created an interface for session configuration, with the latter having
an implementation for all but the session store
- Callers should now use PersistableSessions instead of
RethinkDbSessions, still passing the RethinkDB configuration object
This code set is not ready for prime time, but it's as good as the
combined Persistable/RethinkDB packages were prior to this commit.  A
new NuGet package will be generated once v0.8.6 is ready.