Rework Relational API
The original design had been for the consumer to pass a SQL dialect enum value, and we would derive everything from that; however, as .NET Core doesn't support looking up DbProviderFactories by name, the API needs a concrete instance. Since that is required, we can derive the dialect from the type of the factory. This also implements the ConnectionString property properly, getting the string from configuration if we're given a name.